+++ functions/execPath.js
@@ -4,8 +4,10 @@ const MODULE_REQUIRE = 1
     /* built-in */
     
     /* NPM */
+    , noda = require('noda')
     
     /* in-package */
+    , flag = noda.inRequire('util/flag')
     ;
 
 /**
@@ -13,6 +15,8 @@ const MODULE_REQUIRE = 1
  * @return {string}
  */
 function execPath() {
+    flag('allow-read');
+    
     return process.execPath;
 }