+++ functions/link.js
@@ -5,8 +5,10 @@ const MODULE_REQUIRE = 1
     , fs = require('fs')
     
     /* NPM */
+    , noda = require('noda')
     
     /* in-package */
+    , flag = noda.inRequire('util/flag')
     ;
 
 /**
@@ -14,7 +16,9 @@ const MODULE_REQUIRE = 1
  * @param {string} oldname
  * @param {string} newname
  */
-function link(oldname, newname) {    
+function link(oldname, newname) {
+    flag('allow-write');
+
     return new Promise((resolve, reject) => {
         fs.link(oldname, newname, err => {
             err ? reject(err) : resolve();