+++ index.js
@@ -11,4 +11,14 @@ const MODULE_REQUIRE = 1
 const Variables = require('./variables');
 const ErrorKind = require('./ErrorKind');
 const Functions = require('./functions')
-module.exports = Object.assign({ ErrorKind }, Variables, Functions);
+
+const Deno = { 
+    /**
+     * @deprecated since 0.1.0
+     */
+    ErrorKind,
+
+    ...Variables, 
+    ...Functions,
+};
+module.exports = Deno;